backend/requirements.txt
# FastAPI core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis cache
redis==5.0.1

# OCR
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
opencv-python-headless==4.9.0.80

# Web scraping
playwright==1.41.2
beautifulsoup4==4.12.3
httpx==0.26.0
lxml==5.1.0

# AI
openai==1.12.0
anthropic==0.18.1

# Barcode
pyzbar==0.1.9

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1

# PDF generation
reportlab==4.1.0
weasyprint==61.0

# Data processing
pandas==2.2.0
numpy==1.26.4
